+++ Exams/Abiturient.cpp
@@ -1,0 +1,1 @@
+#include "Abiturient.h"
