+++ Exams/Exams.cpp
@@ -1,5 +1,5 @@
 // Exams.cpp : Этот файл содержит функцию "main". Здесь начинается и заканчивается выполнение программы.
-
+#include "Rating.h"
 #define _USE_MATH_DEFINES
 #include <iostream>
 #include <iomanip>
@@ -8,91 +8,6 @@
 #include <sstream>
 
 using namespace std;
-
-class Abiturient
-{
-private:
-    class Rating
-    {
-    public:
-        Rating() // конструктор
-        {
-            sz_students = 0;
-            students = nullptr;
-        }
-        Rating(int sz) // перегруженный конструктор
-        {
-            sz_students = sz;
-            students = new Abiturient[sz];
-        }
-        ~Rating() { } // деструктор
-        void task_1(const int sz, const string filename, string& s_z);
-        void get_data(const string filename, const int sz) // чтение и запись из файла
-        {
-            ifstream file_in(filename);
-            string surname, name, patronymic;
-            for (int i = 0; i < sz; i++)
-            {
-                file_in >> surname >> name >> patronymic;
-                students[i].set_st(surname, name, patronymic);
-                students[i].show();
-
-            }
-            file_in.close();
-        };
-    private:
-        Abiturient* students;
-        int sz_students;
-    };
-
-private:
-    string surname;
-    string name;
-    string patronymic;
-public:
-    Abiturient() // конструктор
-    {
-        surname = "nosurname";
-        name = "noname";
-        patronymic = "nopatronymic";
-    }
-    Abiturient(string sn, string n, string p) // перегруженный конструктор
-    {
-        surname = sn;
-        name = n;
-        patronymic = p;
-    }
-    ~Abiturient() { } // деструктор
-    void set_st(string sn, string n, string p)
-    {
-        surname = sn;
-        name = n;
-        patronymic = p;
-    }
-    const string& get_surname() const
-    {
-        return surname;
-    }
-
-    const string& get_name() const
-    {
-        return name;
-    }
-
-    const string& get_patronymic() const
-    {
-        return patronymic;
-    }
-
-    void show()
-    {
-        cout << surname << " " << name << " " << patronymic << endl;
-    }
-    void show_f(string& s_z);
-    void show_1(const string filename, const string sz);
-};
-
-//---------------------------------------------------------------------//
 
 void clear(const string filename);
 void get_data_1(const string filename, int& sz, bool& u, int& y);
@@ -109,7 +24,7 @@ void Abiturient::show_1(const string filename, const string s_z) // запись
     out.close();
 }
 
-void Abiturient::Rating::task_1(const int sz, const string filename, string& s_z)
+void Rating::task_1(const int sz, const string filename, string& s_z)
 {
     string user;
     cin >> user;
@@ -171,7 +86,7 @@ int main() {
 
     setlocale(LC_ALL, "");
 
-    string file ="Файл для хранения данных.txt";
+    string file = "Файл для хранения данных.txt";
     string username;
 
     bool u = false;
@@ -181,7 +96,7 @@ int main() {
     //cout << "Введите название файла из которого будет чтение:" << endl;
     //cin >> file;
     get_data_1(file, sz, u, y);
-    Abiturient::Rating abiturient(sz);
+    Rating abiturient(sz);
 
     unsigned int a = 0;
     do {
@@ -199,7 +114,7 @@ int main() {
             }
             else if ((y == 1) && (u == false))
             {
-                cout << "Ошибка программы..." <<  endl;
+                cout << "Ошибка программы..." << endl;
                 return 0;
             }
