+++ Exams/Exams.vcxproj
@@ -139,7 +139,13 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Abiturient.cpp" />
     <ClCompile Include="Exams.cpp" />
+    <ClCompile Include="Rating.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Abiturient.h" />
+    <ClInclude Include="Rating.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
