+++ Exams/Rating.cpp
@@ -1,0 +1,1 @@
+#include "Rating.h"
